+++ CHANGELOG.md
@@ -118,3 +118,4 @@
 - [2026-04-04 15:13] feat(core): add async core stream provider
 - [2026-04-04 15:19] feat(navigation): add async navigation stream provider
 - [2026-04-04 16:12] test(storage): implement widget integration tests for storage
+- [2026-04-04 16:30] style(core): adjust design tokens, typography, and contrast in core